backend/requirements.txt
# Core dependencies
playwright==1.49.1
python-dotenv==1.0.1
notion-client==2.2.1

# YouTube API
google-api-python-client==2.158.0

# Image processing
opencv-python==4.10.0.84
numpy==2.2.1
Pillow==11.1.0

# Progress bars
tqdm==4.67.1

# Additional utilities
aiofiles==24.1.0

# HTTP requests
requests==2.31.0

# AI
openai==1.59.3
